requirements.txt
python-telegram-bot[webhooks]==20.6
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==1.10.9
jinja2==3.1.3
